requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
requests==2.32.3
python-dotenv==1.0.1
